Combined upc/derived processing, split processing/database functionality
